Instantiating model in repository
