+++ src/data/auth/interfaces/index.ts
@@ -1,0 +1,3 @@
+export * from 'src/data/auth/interfaces/login-data.interface';
+export * from 'src/data/auth/interfaces/login-response.interface';
+export * from 'src/data/auth/interfaces/new-access-token-response.interface';
