+++ src/data/auth/interfaces/login-data.interface.ts
@@ -1,0 +1,4 @@
+export interface LoginData {
+  email: string;
+  password: string;
+}
