+++ src/data/auth/interfaces/login-response.interface.ts
@@ -1,0 +1,5 @@
+/* eslint-disable camelcase */
+
+export interface LoginResponse {
+  access_token: string;
+}
