+++ src/data/auth/interfaces/new-access-token-response.interface.ts
@@ -1,0 +1,3 @@
+import { LoginResponse } from 'src/data/auth/interfaces';
+
+export type NewAccessTokenResponse = LoginResponse;
