+++ src/models/auth/access-token.model.ts
@@ -1,0 +1,7 @@
+export default class AccessTokenModel {
+  public accessToken: string;
+
+  constructor(accessToken: string) {
+    this.accessToken = accessToken;
+  }
+}
